+++ src/Dosepy/app.py
@@ -8,16 +8,19 @@ from PySide6.QtWidgets import (
     QTabWidget,
 )
 
+# Import app views
 from .app_components.calibration_widget import CalibrationWidget
 from .app_components.tiff2dose_widget import Tiff2DoseWidget
-from .app_controller import DosepyController
+# Import app controllers
+from .app_controller import CalibrationController, Tiff2DoseController
+# Import app model
 from .app_model import Model
 
 class MainWindow(QWidget):
     def __init__(self):
         super().__init__()
 
-        self.setWindowTitle("Film dosimetry")
+        self.setWindowTitle("Dosepy")
 
         tab_layout = QVBoxLayout()
         self.setLayout(tab_layout)
@@ -27,7 +30,7 @@ class MainWindow(QWidget):
         tabs.addTab(self.cal_widget, "Calibration")
         self.tif_widget = Tiff2DoseWidget()
         tabs.addTab(self.tif_widget, "Film2Dose")
-        tabs.addTab(QLabel("In progress..."), "Analysis")
+        #tabs.addTab(QLabel("In progress..."), "Analysis")
         tab_layout.addWidget(tabs)
 
 '''
@@ -47,7 +50,8 @@ app = QApplication(sys.argv)
 
 root_window = MainWindow()
 dosepy_model = Model()
-dosepy_controller = DosepyController(model=dosepy_model, view=root_window)
+dosepy_calibration_controller = CalibrationController(model=dosepy_model, view=root_window)
+dosepy_tiff2dose_controller = Tiff2DoseController(model=dosepy_model, view=root_window)
 root_window.show()
 
 sys.exit(app.exec())